+++ jetty-http/src/main/java/org/eclipse/jetty/http/MultiPartFormDataCompliance.java
@@ -16,7 +16,7 @@
 //  ========================================================================
 //
 
-package org.eclipse.jetty.server;
+package org.eclipse.jetty.http;
 
 /**
  * The compliance level for parsing <code>multiPart/form-data</code>
@@ -27,7 +27,7 @@ public enum MultiPartFormDataCompliance
     /**
      * Legacy <code>multiPart/form-data</code> parsing which is slow but forgiving.
      * It will accept non compliant preambles and inconsistent line termination.
-     * @see org.eclipse.jetty.util.MultiPartInputStreamParser
+     * @see org.eclipse.jetty.http.internal.MultiPartInputStreamParser
      */
     LEGACY, 
     /**
